Update medallia to v 4.2.0
