+++ packages/nativescript-medallia/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@essent/nativescript-medallia",
-  "version": "4.1.1",
+  "version": "4.2.0",
   "description": "Nativescript plugin for medallia integration.",
   "main": "index",
   "typings": "index.d.ts",
