+++ packages/nativescript-medallia/platforms/ios/MedalliaDigitalSDK.xcframework/ios-arm64/MedalliaDigitalSDK.framework/Modules/module.modulemap
@@ -1,0 +1,4 @@
+framework module MedalliaDigitalSDK {
+  header "MedalliaDigitalSDK-Swift.h"
+  requires objc
+}
