--- packages/nativescript-medallia/platforms/ios/MedalliaDigitalSDK.xcframework/ios-arm64_armv7/MedalliaDigitalSDK.framework/Modules/module.modulemap
@@ -1,4 +1,0 @@
-framework module MedalliaDigitalSDK {
-  header "MedalliaDigitalSDK-Swift.h"
-  requires objc
-}
