--- packages/nativescript-medallia/platforms/ios/MedalliaDigitalSDK.xcframework/ios-arm64_i386_x86_64-simulator/MedalliaDigitalSDK.framework/Modules/module.modulemap
@@ -1,4 +1,0 @@
-framework module MedalliaDigitalSDK {
-  header "MedalliaDigitalSDK-Swift.h"
-  requires objc
-}
