--- packages/nativescript-medallia/platforms/ios/MedalliaDigitalSDK.xcframework/ios-arm64_i386_x86_64-simulator/MedalliaDigitalSDK.framework/_CodeSignature/CodeResources
@@ -1,297 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Assets.car</key>
-		<data>
-		y9x3NHKiJYmCV6fd7DFTKmKN1As=
-		</data>
-		<key>Headers/MedalliaDigitalSDK-Swift.h</key>
-		<data>
-		xnNonjPIYIgFUyKpDEwAH6U7zrg=
-		</data>
-		<key>Info.plist</key>
-		<data>
-		ZwevFCuc4kwJtV8DvPgSJosrA8E=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
-		<data>
-		c5vJvlM0g3BsW4P/f+KvAXdi8XQ=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
-		<data>
-		IwQzvm4wZIFU8gO48YNGI2kItSk=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
-		<data>
-		Qn8UR6FOUzZrONRSMl4IGsAz+T4=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/i386-apple-ios-simulator.swiftdoc</key>
-		<data>
-		uL+ef7imDLA0/DL0twtljsXwtho=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/i386-apple-ios-simulator.swiftinterface</key>
-		<data>
-		zOA+P2+nYtKbOsZGSa093ZU1cQw=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/i386-apple-ios-simulator.swiftmodule</key>
-		<data>
-		S4rY51qlrbgt383O56bOJMp8l4U=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<data>
-		l4jBLS7QPbsP3Z5P3eeY9nthBP0=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<data>
-		Gj9ZY0ECDeR8lQ+HhF1nEP/O4cw=
-		</data>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<data>
-		a1APEhexeFey65CHyFQG2cY69XI=
-		</data>
-		<key>Modules/module.modulemap</key>
-		<data>
-		aJ7QCkTgHQh91XPW/tOg6T2TXHQ=
-		</data>
-		<key>stripArchitechtures.sh</key>
-		<data>
-		AQGtHOPcQneTCmSkBCyvHX8DjPk=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Assets.car</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			y9x3NHKiJYmCV6fd7DFTKmKN1As=
-			</data>
-			<key>hash2</key>
-			<data>
-			+sSOEPDGw2EqgSro0FbE8kWcfOWJaoVIPa0ZvB8L/wo=
-			</data>
-		</dict>
-		<key>Headers/MedalliaDigitalSDK-Swift.h</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			xnNonjPIYIgFUyKpDEwAH6U7zrg=
-			</data>
-			<key>hash2</key>
-			<data>
-			6d3rkPKNXB5LBGbiJyom8I+ufgjzTxMRNwiPVs2niME=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			c5vJvlM0g3BsW4P/f+KvAXdi8XQ=
-			</data>
-			<key>hash2</key>
-			<data>
-			LQiIDKEJjVSw9pN9sZUr8cqM39vEfPjm/9gybHw4o+E=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			IwQzvm4wZIFU8gO48YNGI2kItSk=
-			</data>
-			<key>hash2</key>
-			<data>
-			nZP+6hchu/AVmjs6PYgalm6JPf7iw8q33GgKkQADptw=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			Qn8UR6FOUzZrONRSMl4IGsAz+T4=
-			</data>
-			<key>hash2</key>
-			<data>
-			Hc176vqQ5ve2F3aiM6Gn1xpDUt/UbAqC7QUdHjEcMTs=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/i386-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			uL+ef7imDLA0/DL0twtljsXwtho=
-			</data>
-			<key>hash2</key>
-			<data>
-			L7Gcf9gWbjg5Y1ua81DkKxeIurOsvw/wGWuTqGBg17c=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/i386-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			zOA+P2+nYtKbOsZGSa093ZU1cQw=
-			</data>
-			<key>hash2</key>
-			<data>
-			Q8FUK2mKc7JnMgZifIgyX8nflhS2euPc5u6ZL8VoWZk=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/i386-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			S4rY51qlrbgt383O56bOJMp8l4U=
-			</data>
-			<key>hash2</key>
-			<data>
-			BHRbhs3FzRk460A3hNpZxYETvyMqAKs+sH1cuUDGFjc=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			l4jBLS7QPbsP3Z5P3eeY9nthBP0=
-			</data>
-			<key>hash2</key>
-			<data>
-			X1bkRser/9qmjvsCir4ipRdFXZ6s1qF4tuFCN96oFQ0=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			Gj9ZY0ECDeR8lQ+HhF1nEP/O4cw=
-			</data>
-			<key>hash2</key>
-			<data>
-			gx6wV7Wk6TmAlQnkEV+ziBfyC48BbSaosZcGTyvwKDQ=
-			</data>
-		</dict>
-		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			a1APEhexeFey65CHyFQG2cY69XI=
-			</data>
-			<key>hash2</key>
-			<data>
-			5rDXMqb6BhppS1JEWETJkwXLhJJ3mNcx+82Keorz5cQ=
-			</data>
-		</dict>
-		<key>Modules/module.modulemap</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			aJ7QCkTgHQh91XPW/tOg6T2TXHQ=
-			</data>
-			<key>hash2</key>
-			<data>
-			ibKxuetA9pf5k7JngoutavfOFSYITbkuc+R7sYAgPa4=
-			</data>
-		</dict>
-		<key>stripArchitechtures.sh</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			AQGtHOPcQneTCmSkBCyvHX8DjPk=
-			</data>
-			<key>hash2</key>
-			<data>
-			N9+MGizcCBqJiCDvzSsGSvjtt/FNNu/ciszGaJEBXbw=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
