+++ packages/nativescript-medallia/platforms/ios/MedalliaDigitalSDK.xcframework/ios-arm64_x86_64-simulator/MedalliaDigitalSDK.framework/_CodeSignature/CodeResources
@@ -1,0 +1,312 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Assets.car</key>
+		<data>
+		BZhVqr6Svpg0e9uqZ8yXR9WTfBY=
+		</data>
+		<key>Headers/MedalliaDigitalSDK-Swift.h</key>
+		<data>
+		s+p5UNaBE/TCJwvZfj6RVwpNo8Y=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		t0L1M4cw2Ld3hUiNt8JSJNybupQ=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
+		<data>
+		1ZTVHV7Gcf/2oxgkPndoaYqk3dQ=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
+		<data>
+		OZcpt1W95/SoD9pUJ4hRHRwIdS8=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		L9oULUPAlEn9ddFSpU+V1YRYIkc=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		OZcpt1W95/SoD9pUJ4hRHRwIdS8=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		aM8310C799ea3gFfwxGOq7SoBDU=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<data>
+		1ZTVHV7Gcf/2oxgkPndoaYqk3dQ=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<data>
+		kldnm2RRXIfaSf9ScoHosuAVuvQ=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		GRggMj+RNgIDptnFlCExIaoJXxA=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		kldnm2RRXIfaSf9ScoHosuAVuvQ=
+		</data>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		ISgOMrKYvVvvJZHkvpZTFleg30o=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		aJ7QCkTgHQh91XPW/tOg6T2TXHQ=
+		</data>
+		<key>stripArchitechtures.sh</key>
+		<data>
+		AQGtHOPcQneTCmSkBCyvHX8DjPk=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Assets.car</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			BZhVqr6Svpg0e9uqZ8yXR9WTfBY=
+			</data>
+			<key>hash2</key>
+			<data>
+			uF9abrda5f9wCYozAayLgj+pHW7NEwOHo4NMKRKLzEo=
+			</data>
+		</dict>
+		<key>Headers/MedalliaDigitalSDK-Swift.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			s+p5UNaBE/TCJwvZfj6RVwpNo8Y=
+			</data>
+			<key>hash2</key>
+			<data>
+			/xNJQO2kcQDDMI97K1kNmw7ycrCJA34IBAodmdl8UY0=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.abi.json</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			1ZTVHV7Gcf/2oxgkPndoaYqk3dQ=
+			</data>
+			<key>hash2</key>
+			<data>
+			mAOPsc1PkAHzb4f3LdWyWQ2tAfuunFETIQTaeZocjgc=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.private.swiftinterface</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			OZcpt1W95/SoD9pUJ4hRHRwIdS8=
+			</data>
+			<key>hash2</key>
+			<data>
+			qTZE3rmmauv/+9cJvT71H9LRPP6Cme6i0mIc4FKqMZ8=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			L9oULUPAlEn9ddFSpU+V1YRYIkc=
+			</data>
+			<key>hash2</key>
+			<data>
+			G6zCixlmCweYmc9Qh16Dh8TKtjc92u8d81RdIdm2Czs=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			OZcpt1W95/SoD9pUJ4hRHRwIdS8=
+			</data>
+			<key>hash2</key>
+			<data>
+			qTZE3rmmauv/+9cJvT71H9LRPP6Cme6i0mIc4FKqMZ8=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/arm64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			aM8310C799ea3gFfwxGOq7SoBDU=
+			</data>
+			<key>hash2</key>
+			<data>
+			6Y61DHeFpSyn9rPVmSiHjCnP0DQG3if68/nG7OHMw5g=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.abi.json</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			1ZTVHV7Gcf/2oxgkPndoaYqk3dQ=
+			</data>
+			<key>hash2</key>
+			<data>
+			mAOPsc1PkAHzb4f3LdWyWQ2tAfuunFETIQTaeZocjgc=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.private.swiftinterface</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			kldnm2RRXIfaSf9ScoHosuAVuvQ=
+			</data>
+			<key>hash2</key>
+			<data>
+			v3S3p0m2nl8RLOxh8ZQTkTt8BemUw2DwDFaKfef8ruE=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			GRggMj+RNgIDptnFlCExIaoJXxA=
+			</data>
+			<key>hash2</key>
+			<data>
+			jD1cRjadBAaZAxfNUeBVO3mj34eZVgEg64MN8jcz99o=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			kldnm2RRXIfaSf9ScoHosuAVuvQ=
+			</data>
+			<key>hash2</key>
+			<data>
+			v3S3p0m2nl8RLOxh8ZQTkTt8BemUw2DwDFaKfef8ruE=
+			</data>
+		</dict>
+		<key>Modules/MedalliaDigitalSDK.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			ISgOMrKYvVvvJZHkvpZTFleg30o=
+			</data>
+			<key>hash2</key>
+			<data>
+			Y5JWMJHXRZZHtYYtLIHgvZCSFGC70GbwaocKWuOqpOI=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			aJ7QCkTgHQh91XPW/tOg6T2TXHQ=
+			</data>
+			<key>hash2</key>
+			<data>
+			ibKxuetA9pf5k7JngoutavfOFSYITbkuc+R7sYAgPa4=
+			</data>
+		</dict>
+		<key>stripArchitechtures.sh</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			AQGtHOPcQneTCmSkBCyvHX8DjPk=
+			</data>
+			<key>hash2</key>
+			<data>
+			N9+MGizcCBqJiCDvzSsGSvjtt/FNNu/ciszGaJEBXbw=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
